refactor: icon import 수정
